streaming_ssnhash_pubsub/dlp_trigger_function/requirements.txt
google-cloud-dlp
google-cloud-storage
google-cloud-pubsub
